requirements.txt
numpy==1.24.2
casadi==3.5.6rc2
matplotlib==3.7.1
joblib==1.2.0
PyYAML==6.0
tqdm==4.65.0
scipy==1.10.1
pandas==1.5.3
shapely==1.8.5